pyproject.toml
[project]
name = "chl-mcp-server"
version = "0.2.0"
description = "Curated Heuristic Loop MCP Server for experience management with semantic search"
requires-python = ">=3.10,<3.12"  # Torch 2.0.x wheels are only for Python 3.10–3.11
dependencies = [
    "fastmcp>=0.3.0",
    "sqlalchemy>=2.0.0",
    "numpy>=1.24.0,<2.0.0",
    "gspread>=5.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
]
ml = [
    # Using llama-cpp-python for GGUF quantized embeddings
    "llama-cpp-python>=0.2.0",
    "faiss-cpu>=1.7.4",
    "huggingface-hub>=0.20.0",  # For downloading GGUF models
    "sentence-transformers>=3.1.1",
]
test = [
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
